pyproject.toml
[tool.poetry]
name = "django-utils"
version = "2023.1.0"
description = "A collection of utils used in our Django based web applications."
authors = ["4teamwork AG"]
readme = "README.md"
packages = [{include = "django_utils"}]

[tool.poetry.dependencies]
python = ">=3.9, <4"
django-configurations = "^2.4.1"
django-extensions = "^3.2.1"
sentry-sdk = "^1.21.0"
djangorestframework = "^3.14.0"
django = "^3"
django-sendfile2 = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
exclude = '''
(
  /(
      \.git
    | \.venv
    | \.eggs
  )/
)
'''

[tool.isort]
force_alphabetical_sort_within_sections = true
force_single_line = true
from_first = false
known_first_party = "django_utils"
line_length = 125
lines_after_imports = 2
skip_glob = ".venv,.eggs"
